feat(developer): restructured folders to follow MVC -added CI badges from Travis and Coveralls to README [Delivers #164537581]
