+++ README.md
@@ -1,2 +1,6 @@
 # epic-mail
 web app that helps people exchange messages/information over the internet
+[![Build Status](https://travis-ci.org/ola357/epic-mail.svg?branch=develop)](https://travis-ci.org/ola357/epic-mail)
+
+
+[![Coverage Status](https://coveralls.io/repos/github/ola357/epic-mail/badge.svg)](https://coveralls.io/github/ola357/epic-mail)
